+++ src/mesh/GenMesh.cpp
@@ -3,23 +3,26 @@
 
 namespace tndm {
 
-template<>
-std::array<GenMesh<1>::simplex_t,1> GenMesh<1>::tessellate(std::array<int,2> const& vertGIDs, bool) {
+template <>
+std::array<GenMesh<1>::simplex_t, 1> GenMesh<1>::tessellate(std::array<uint64_t, 2> const& vertGIDs,
+                                                            bool) {
     std::array<simplex_t,1> result;
     result[0] = vertGIDs;
     return result;
 }
 
-template<>
-std::array<GenMesh<2>::simplex_t,2> GenMesh<2>::tessellate(std::array<int,4> const& vertGIDs, bool isOdd) {
+template <>
+std::array<GenMesh<2>::simplex_t, 2> GenMesh<2>::tessellate(std::array<uint64_t, 4> const& vertGIDs,
+                                                            bool isOdd) {
     std::array<simplex_t,2> result;
     result[0] = {vertGIDs[0], vertGIDs[1], vertGIDs[2]};
     result[1] = {vertGIDs[1], vertGIDs[3], vertGIDs[2]};
     return result;
 }
 
-template<>
-std::array<GenMesh<3>::simplex_t,5> GenMesh<3>::tessellate(std::array<int,8> const& vertGIDs, bool isOdd) {
+template <>
+std::array<GenMesh<3>::simplex_t, 5> GenMesh<3>::tessellate(std::array<uint64_t, 8> const& vertGIDs,
+                                                            bool isOdd) {
     std::array<simplex_t,5> result;
     if (isOdd) {
         result[0] = {vertGIDs[0], vertGIDs[1], vertGIDs[2], vertGIDs[4]};
@@ -36,7 +39,6 @@ std::array<GenMesh<3>::simplex_t,5> GenMesh<3>::tessellate(std::array<int,8> con
     }
     return result;
 }
-
 }
